+++ src/main/java/com/tagnumelite/projecteintegration/addons/AppliedEnergisticsAddon.java
@@ -22,20 +22,20 @@
 
 package com.tagnumelite.projecteintegration.addons;
 
+import appeng.core.definitions.AEBlocks;
 import appeng.core.definitions.AEItems;
 import appeng.recipes.AERecipeTypes;
 import appeng.recipes.handlers.ChargerRecipe;
 import appeng.recipes.handlers.InscriberProcessType;
 import appeng.recipes.handlers.InscriberRecipe;
+import appeng.recipes.transform.TransformRecipe;
 import com.tagnumelite.projecteintegration.api.conversion.AConversionProvider;
 import com.tagnumelite.projecteintegration.api.conversion.ConversionProvider;
 import com.tagnumelite.projecteintegration.api.recipe.ARecipeTypeMapper;
 import moze_intel.projecte.api.data.CustomConversionBuilder;
 import moze_intel.projecte.api.mapper.recipe.RecipeTypeMapper;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeType;
-import net.neoforged.neoforge.common.Tags;
 
 import java.util.Collections;
 import java.util.List;
@@ -60,6 +60,8 @@ public class AppliedEnergisticsAddon {
         }
     }
 
+    // TODO: EntropyMapper
+
     @RecipeTypeMapper(requiredMods = MODID, priority = 1)
     public static class AEInscriberMapper extends ARecipeTypeMapper<InscriberRecipe> {
         @Override
@@ -74,11 +76,24 @@ public class AppliedEnergisticsAddon {
 
         @Override
         protected List<Ingredient> getIngredients(InscriberRecipe recipe) {
-            // Whoops, forgot that inscribing exists.
+            // The Inscribe recipe type doesn't use the top and bottom input
             if (recipe.getProcessType() == InscriberProcessType.INSCRIBE) {
                 return Collections.singletonList(recipe.getMiddleInput());
             }
             return super.getIngredients(recipe);
+        }
+    }
+
+    @RecipeTypeMapper(requiredMods = MODID, priority = 1)
+    public static class AETransformMapper extends ARecipeTypeMapper<TransformRecipe> {
+        @Override
+        public String getName() {
+            return NAME("Transform");
+        }
+
+        @Override
+        public boolean canHandle(RecipeType<?> recipeType) {
+            return recipeType == AERecipeTypes.TRANSFORM;
         }
     }
 
@@ -88,11 +103,7 @@ public class AppliedEnergisticsAddon {
         public void convert(CustomConversionBuilder builder) {
             builder.comment("Set defaults conversions for Applied Energistics")
                     .before(AEItems.CERTUS_QUARTZ_CRYSTAL.get(), 256)
-                    //.conversion(AEItems.CERTUS_QUARTZ_CRYSTAL_CHARGED.get()).ingredient(AEItems.CERTUS_QUARTZ_CRYSTAL.get()).end()
-                    .conversion(AEItems.FLUIX_DUST.get(), 2).ingredient(AEItems.CERTUS_QUARTZ_CRYSTAL_CHARGED.get()).ingredient(Tags.Items.DUSTS_REDSTONE).ingredient(Items.QUARTZ).end();
-            //.conversion(Items.QUARTZ).ingredient(NETHER_QUARTZ_SEED).end()
-            //.conversion(FLUIX_CRYSTAL.get()).ingredient(FLUIX_CRYSTAL_SEED.get()).end()
-            //.conversion(CERTUS_QUARTZ_CRYSTAL.get()).ingredient(CERTUS_CRYSTAL_SEED.get()).end();
+                    .before(AEBlocks.SKY_STONE_BLOCK, 16);
         }
     }
 }